+++ tests/strategies/test_ddp_strategy.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
+from datetime import timedelta
 from unittest import mock
 
 import pytest
@@ -226,3 +227,24 @@ def test_configure_launcher_create_processes_externally():
     assert ddp_strategy.launcher is None
     ddp_strategy._configure_launcher()
     assert ddp_strategy.launcher is None
+    
+
+@RunIf(min_cuda_gpus=1)
+@mock.patch("torch.distributed.init_process_group")
+def test_ddp_strategy_set_timeout(mock_init_process_group):
+    """Tests with ddp strategy."""
+    test_timedelta = timedelta(seconds=30)
+    model = BoringModel()
+    ddp_strategy = DDPStrategy(timeout=test_timedelta)
+    trainer = Trainer(
+        max_epochs=1,
+        strategy=ddp_strategy,
+    )
+    # test wrap the model if fitting
+    trainer.state.fn = TrainerFn.FITTING
+    trainer.strategy.connect(model)
+    trainer.lightning_module.trainer = trainer
+    trainer.strategy.setup_environment()
+    mock_init_process_group.assert_called_with(timeout=test_timedelta)
+
+